Migrated the rest of the streams ITs off the old Agrona-based scripts.
