+++ src/test/java/org/reaktivity/nukleus/tcp/internal/SocketChannelHelper.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.nukleus.tcp.internal.streams;
+package org.reaktivity.nukleus.tcp.internal;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
