+++ src/test/java/org/reaktivity/nukleus/tcp/internal/streams/rfc793/ClientPartialWriteIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.nukleus.tcp.internal.streams;
+package org.reaktivity.nukleus.tcp.internal.streams.rfc793;
 
 import static java.net.StandardSocketOptions.SO_REUSEADDR;
 import static java.nio.charset.StandardCharsets.UTF_8;
@@ -24,7 +24,7 @@ import static java.util.stream.IntStream.of;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.rules.RuleChain.outerRule;
-import static org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.ALL;
+import static org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.ALL;
 import static org.reaktivity.nukleus.tcp.internal.stream.WriteStream.WRITE_SPIN_COUNT;
 
 import java.net.InetSocketAddress;
@@ -44,11 +44,11 @@ import org.junit.rules.Timeout;
 import org.junit.runner.RunWith;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.HandleWriteHelper;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.ProcessDataHelper;
 import org.reaktivity.nukleus.tcp.internal.TcpController;
-import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.HandleWriteHelper;
-import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.ProcessDataHelper;
 import org.reaktivity.reaktor.test.ReaktorRule;
-import org.reaktivity.specification.nukleus.NukleusRule;
 
 @RunWith(org.jboss.byteman.contrib.bmunit.BMUnitRunner.class)
 @BMUnitConfig(loadDirectory="src/test/resources")
@@ -57,7 +57,8 @@ public class ClientPartialWriteIT
 {
     private final K3poRule k3po = new K3poRule()
         .addScriptRoot("route", "org/reaktivity/specification/nukleus/tcp/control/route")
-        .addScriptRoot("streams", "org/reaktivity/specification/nukleus/tcp/streams");
+        .addScriptRoot("server", "org/reaktivity/specification/tcp/rfc793")
+        .addScriptRoot("client", "org/reaktivity/specification/nukleus/tcp/streams/rfc793");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
@@ -67,55 +68,55 @@ public class ClientPartialWriteIT
         .directory("target/nukleus-itests")
         .commandBufferCapacity(1024)
         .responseBufferCapacity(1024)
-        .counterValuesBufferCapacity(1024);
-
-    private final NukleusRule file = new NukleusRule()
-            .directory("target/nukleus-itests")
-            .streams("tcp", "source#partition")
-            .streams("source", "tcp#source");
+        .counterValuesBufferCapacity(1024)
+        .clean();
 
     @Rule
-    public final TestRule chain = outerRule(SocketChannelHelper.RULE).around(file).around(reaktor).around(k3po).around(timeout);
+    public final TestRule chain = outerRule(SocketChannelHelper.RULE).around(reaktor).around(k3po).around(timeout);
 
     @Test
     @Specification({
         "${route}/client/controller",
-        "${streams}/client.sent.data/client/source"
+        "${client}/client.sent.data/client",
+        "${server}/client.sent.data/server"
     })
     public void shouldSpinWrite() throws Exception
     {
         ProcessDataHelper.fragmentWrites(generate(() -> 0).limit(WRITE_SPIN_COUNT - 1));
         HandleWriteHelper.fragmentWrites(generate(() -> 0));
-        shouldReceiveClientSentData("client data");
+        k3po.finish();
     }
 
     @Test
     @Specification({
         "${route}/client/controller",
-        "${streams}/client.sent.data/client/source"
+        "${client}/client.sent.data/client",
+        "${server}/client.sent.data/server"
     })
     public void shouldFinishWriteWhenSocketIsWritableAgain() throws Exception
     {
         ProcessDataHelper.fragmentWrites(IntStream.of(5));
-        shouldReceiveClientSentData("client data");
+        k3po.finish();
     }
 
     @Test
     @Specification({
         "${route}/client/controller",
-        "${streams}/client.sent.data/client/source"
+        "${client}/client.sent.data/client",
+        "${server}/client.sent.data/server"
     })
     public void shouldHandleMultiplePartialWrites() throws Exception
     {
         ProcessDataHelper.fragmentWrites(IntStream.of(2));
         HandleWriteHelper.fragmentWrites(IntStream.of(3, 1));
-        shouldReceiveClientSentData("client data");
+        k3po.finish();
     }
 
     @Test
     @Specification({
         "${route}/client/controller",
-        "${streams}/client.sent.data.multiple.frames/client/source"
+        "${client}/client.sent.data.multiple.frames/client",
+        "${server}/client.sent.data.multiple.frames/server"
     })
     public void shouldWriteWhenMoreDataArrivesWhileAwaitingSocketWritable() throws Exception
     {
@@ -126,14 +127,13 @@ public class ClientPartialWriteIT
 
         ProcessDataHelper.fragmentWrites(concat(of(5), generate(() -> finishWrite.getAndSet(true) ? 0 : 0)));
         HandleWriteHelper.fragmentWrites(generate(() -> finishWrite.get() ? ALL : 0));
-
-        shouldReceiveClientSentData("client data 1client data 2");
+        k3po.finish();
     }
 
     @Test
     @Specification({
         "${route}/client/controller",
-        "${streams}/client.sent.data.then.end/client/source"
+        "${client}/client.sent.data.then.end/client"
     })
     public void shouldHandleEndOfStreamWithPendingWrite() throws Exception
     {
@@ -151,7 +151,6 @@ public class ClientPartialWriteIT
 
             try (SocketChannel channel = server.accept())
             {
-                k3po.notifyBarrier("CONNECTED_CLIENT");
                 k3po.awaitBarrier("END_WRITTEN");
                 endWritten.set(true);
 
@@ -177,110 +176,6 @@ public class ClientPartialWriteIT
                 }
 
                 assertTrue("Stream was not closed", closed);
-
-                k3po.finish();
-            }
-        }
-    }
-
-    @Test
-    @Specification({
-        "${route}/client/controller",
-        "${streams}/client.sent.data.after.end/client/source"
-    })
-    public void shouldResetIfDataReceivedAfterEndOfStreamWithPendingWrite() throws Exception
-    {
-        ProcessDataHelper.fragmentWrites(IntStream.of(6));
-        AtomicBoolean resetReceived = new AtomicBoolean(false);
-        HandleWriteHelper.fragmentWrites(generate(() -> resetReceived.get() ? ALL : 0));
-
-        try (ServerSocketChannel server = ServerSocketChannel.open())
-        {
-            server.setOption(SO_REUSEADDR, true);
-            server.bind(new InetSocketAddress("127.0.0.1", 0x1f90));
-
-            k3po.start();
-            k3po.awaitBarrier("ROUTED_CLIENT");
-
-            try (SocketChannel channel = server.accept())
-            {
-                k3po.notifyBarrier("CONNECTED_CLIENT");
-
-                k3po.awaitBarrier("RESET_CLIENT");
-                resetReceived.set(true);
-
-                ByteBuffer buf = ByteBuffer.allocate("client data".length() + 10);
-                boolean closed = false;
-                do
-                {
-                    int len = channel.read(buf);
-                    if (len == -1)
-                    {
-                        closed = true;
-                        break;
-                    }
-                } while (buf.position() < "client data".length());
-                buf.flip();
-
-                assertEquals("client data", UTF_8.decode(buf).toString());
-
-                if (!closed)
-                {
-                    buf.rewind();
-                    closed = (channel.read(buf) == -1);
-                }
-
-                assertTrue("Stream was not closed", closed);
-
-                k3po.finish();
-            }
-        }
-    }
-
-    private void shouldReceiveClientSentData(String expectedData) throws Exception
-    {
-        shouldReceiveClientSentData(expectedData, false);
-    }
-
-    private void shouldReceiveClientSentData(String expectedData, boolean expectStreamClosed) throws Exception
-    {
-        try (ServerSocketChannel server = ServerSocketChannel.open())
-        {
-            server.setOption(SO_REUSEADDR, true);
-            server.bind(new InetSocketAddress("127.0.0.1", 0x1f90));
-
-            k3po.start();
-            k3po.awaitBarrier("ROUTED_CLIENT");
-
-            try (SocketChannel channel = server.accept())
-            {
-                k3po.notifyBarrier("CONNECTED_CLIENT");
-
-                ByteBuffer buf = ByteBuffer.allocate(expectedData.length() + 10);
-                boolean closed = false;
-                do
-                {
-                    int len = channel.read(buf);
-                    if (len == -1)
-                    {
-                        closed = true;
-                        break;
-                    }
-                } while (buf.position() < expectedData.length());
-                buf.flip();
-
-                assertEquals(expectedData, UTF_8.decode(buf).toString());
-
-                if (expectStreamClosed)
-                {
-                    if (!closed)
-                    {
-                        buf.rewind();
-                        closed = (channel.read(buf) == -1);
-                    }
-
-                    assertTrue("Stream was not closed", closed);
-                }
 
                 k3po.finish();
             }
