+++ src/test/java/org/reaktivity/nukleus/tcp/internal/streams/rfc793/ClientPartialWriteLimitsIT.java
@@ -13,20 +13,21 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.nukleus.tcp.internal.streams;
+package org.reaktivity.nukleus.tcp.internal.streams.rfc793;
 
+import static java.net.StandardSocketOptions.SO_REUSEADDR;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static java.util.stream.IntStream.concat;
 import static java.util.stream.IntStream.generate;
 import static java.util.stream.IntStream.of;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.rules.RuleChain.outerRule;
-import static org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.ALL;
+import static org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.ALL;
 
 import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
+import java.nio.channels.ServerSocketChannel;
 import java.nio.channels.SocketChannel;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -41,13 +42,13 @@ import org.junit.rules.Timeout;
 import org.junit.runner.RunWith;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.HandleWriteHelper;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.ProcessDataHelper;
 import org.reaktivity.nukleus.tcp.internal.TcpController;
 import org.reaktivity.nukleus.tcp.internal.TcpCountersRule;
-import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.HandleWriteHelper;
-import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.ProcessDataHelper;
 import org.reaktivity.reaktor.internal.ReaktorConfiguration;
 import org.reaktivity.reaktor.test.ReaktorRule;
-import org.reaktivity.specification.nukleus.NukleusRule;
 
 /**
  * Tests the handling of capacity exceeded conditions in the context of incomplete writes
@@ -55,11 +56,12 @@ import org.reaktivity.specification.nukleus.NukleusRule;
 @RunWith(org.jboss.byteman.contrib.bmunit.BMUnitRunner.class)
 @BMUnitConfig(loadDirectory="src/test/resources")
 @BMScript(value="SocketChannelHelper.btm")
-public class ServerPartialWriteLimitsIT
+public class ClientPartialWriteLimitsIT
 {
     private final K3poRule k3po = new K3poRule()
         .addScriptRoot("route", "org/reaktivity/specification/nukleus/tcp/control/route")
-        .addScriptRoot("streams", "org/reaktivity/specification/nukleus/tcp/streams");
+        .addScriptRoot("server", "org/reaktivity/specification/tcp/rfc793")
+        .addScriptRoot("client", "org/reaktivity/specification/nukleus/tcp/streams/rfc793");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
@@ -77,19 +79,15 @@ public class ServerPartialWriteLimitsIT
 
     private final TcpCountersRule counters = new TcpCountersRule(reaktor);
 
-    private final NukleusRule file = new NukleusRule()
-            .directory("target/nukleus-itests")
-            .streams("tcp", "target#partition")
-            .streams("target", "tcp#any");
-
     @Rule
     public final TestRule chain = outerRule(SocketChannelHelper.RULE)
-                                  .around(file).around(reaktor).around(counters).around(k3po).around(timeout);
+                                  .around(reaktor).around(counters).around(k3po).around(timeout);
 
     @Test
     @Specification({
-        "${route}/server/controller",
-        "${streams}/server.sent.data.multiple.frames.partial.writes/server/target"
+        "${route}/client/controller",
+        "${client}/client.sent.data.multiple.frames/client",
+        "${server}/client.sent.data.multiple.frames/server"
     })
     public void shouldWriteWhenMoreDataArrivesWhileAwaitingSocketWritableWithoutOverflowingSlot() throws Exception
     {
@@ -98,41 +96,17 @@ public class ServerPartialWriteLimitsIT
                 : dataFramesReceived.get() == 2 ? 6 : ALL));
         HandleWriteHelper.fragmentWrites(generate(() -> dataFramesReceived.get() >= 2 ? ALL : 0));
 
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_SERVER");
-
-        try (SocketChannel channel = SocketChannel.open())
-        {
-            channel.connect(new InetSocketAddress("127.0.0.1", 0x1f90));
-
-            ByteBuffer buf = ByteBuffer.allocate(256);
-            boolean closed = false;
-            do
-            {
-                int len = channel.read(buf);
-                if (len == -1)
-                {
-                    closed = true;
-                    break;
-                }
-            } while (buf.position() < "server data 1".concat("server data 2").length());
-            buf.flip();
-
-            assertFalse(closed);
-            assertEquals("server data 1".concat("server data 2"), UTF_8.decode(buf).toString());
-
-            k3po.finish();
-        }
+        k3po.finish();
 
         assertEquals(1, counters.streams());
-        assertEquals(0, counters.routes());
+        assertEquals(1, counters.routes());
         assertEquals(0, counters.overflows());
     }
 
     @Test
     @Specification({
-        "${route}/server/controller",
-        "${streams}/server.sent.data.multiple.streams.second.was.reset/server/target"
+        "${route}/client/controller",
+        "${client}/client.sent.data.multiple.streams.second.was.reset/client"
     })
     public void shouldResetStreamsExceedingPartialWriteStreamsLimit() throws Exception
     {
@@ -140,52 +114,50 @@ public class ServerPartialWriteLimitsIT
         AtomicBoolean resetReceived = new AtomicBoolean(false);
         HandleWriteHelper.fragmentWrites(generate(() -> resetReceived.get() ? ALL : 0));
 
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_SERVER");
-
-        try (SocketChannel channel1 = SocketChannel.open();
-             SocketChannel channel2 = SocketChannel.open())
+        try (ServerSocketChannel server = ServerSocketChannel.open())
         {
-            channel1.connect(new InetSocketAddress("127.0.0.1", 0x1f90));
-            channel2.connect(new InetSocketAddress("127.0.0.1", 0x1f90));
+            server.setOption(SO_REUSEADDR, true);
+            server.bind(new InetSocketAddress("127.0.0.1", 0x1f90));
 
-            k3po.awaitBarrier("SECOND_STREAM_RESET_RECEIVED");
-            resetReceived.set(true);
+            k3po.start();
+            k3po.awaitBarrier("ROUTED_CLIENT");
 
-            ByteBuffer buf = ByteBuffer.allocate(256);
-            while (buf.position() < 13)
+            try (SocketChannel channel1 = server.accept();
+                 SocketChannel channel2 = server.accept())
             {
-                int len = channel1.read(buf);
-                if (len == -1)
+                k3po.awaitBarrier("CLIENT_TWO_RESET_RECEIVED");
+                resetReceived.set(true);
+
+                ByteBuffer buf = ByteBuffer.allocate(256);
+                while (buf.position() < 13)
                 {
-                    break;
+                    int len = channel1.read(buf);
+                    assert (len != -1);
                 }
-            }
-            buf.flip();
+                buf.flip();
+                assertEquals("client data 1", UTF_8.decode(buf).toString());
 
-            assertEquals("server data 1", UTF_8.decode(buf).toString());
-
-            int len = 0;
-            buf.rewind();
-            while (buf.position() < 13)
-            {
-                len = channel2.read(buf);
-                if (len == -1)
+                buf.rewind();
+                int len = 0;
+                while (buf.position() < 13)
                 {
-                    break;
+                    len = channel2.read(buf);
+                    if (len == -1)
+                    {
+                        break;
+                    }
                 }
+                buf.flip();
+
+                assertEquals(0, buf.remaining());
+                assertEquals(-1, len);
+
+                k3po.finish();
             }
-            buf.flip();
-
-            assertEquals(0, buf.remaining());
-            assertEquals(-1, len);
-
-            k3po.finish();
         }
 
-        assertEquals(2, counters.streams());
-        assertEquals(0, counters.routes());
+        assertEquals(1, counters.routes());
         assertEquals(1, counters.overflows());
+        assertEquals(2, counters.streams());
     }
 }
-
