+++ src/test/java/org/reaktivity/nukleus/tcp/internal/streams/rfc793/ServerPartialWriteIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.nukleus.tcp.internal.streams;
+package org.reaktivity.nukleus.tcp.internal.streams.rfc793;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.concurrent.TimeUnit.SECONDS;
@@ -23,7 +23,7 @@ import static java.util.stream.IntStream.of;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.rules.RuleChain.outerRule;
-import static org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.ALL;
+import static org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.ALL;
 import static org.reaktivity.nukleus.tcp.internal.stream.WriteStream.WRITE_SPIN_COUNT;
 
 import java.net.InetSocketAddress;
@@ -42,11 +42,11 @@ import org.junit.rules.Timeout;
 import org.junit.runner.RunWith;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.HandleWriteHelper;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.ProcessDataHelper;
 import org.reaktivity.nukleus.tcp.internal.TcpController;
-import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.HandleWriteHelper;
-import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.ProcessDataHelper;
 import org.reaktivity.reaktor.test.ReaktorRule;
-import org.reaktivity.specification.nukleus.NukleusRule;
 
 /**
  * This test verifies the handling of incomplete writes, when attempts to write data to a socket channel
@@ -62,8 +62,8 @@ public class ServerPartialWriteIT
 {
     private final K3poRule k3po = new K3poRule()
         .addScriptRoot("route", "org/reaktivity/specification/nukleus/tcp/control/route")
-        .addScriptRoot("streams", "org/reaktivity/specification/nukleus/tcp/streams");
-
+        .addScriptRoot("client", "org/reaktivity/specification/tcp/rfc793")
+        .addScriptRoot("server", "org/reaktivity/specification/nukleus/tcp/streams/rfc793");
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
     private final ReaktorRule reaktor = new ReaktorRule()
@@ -72,55 +72,55 @@ public class ServerPartialWriteIT
         .directory("target/nukleus-itests")
         .commandBufferCapacity(1024)
         .responseBufferCapacity(1024)
-        .counterValuesBufferCapacity(1024);
-
-    private final NukleusRule file = new NukleusRule()
-            .directory("target/nukleus-itests")
-            .streams("tcp", "target#partition")
-            .streams("target", "tcp#any");
+        .counterValuesBufferCapacity(1024)
+        .clean();
 
     @Rule
-    public final TestRule chain = outerRule(SocketChannelHelper.RULE).around(file).around(reaktor).around(k3po).around(timeout);
+    public final TestRule chain = outerRule(SocketChannelHelper.RULE).around(reaktor).around(k3po).around(timeout);
 
     @Test
     @Specification({
         "${route}/server/controller",
-        "${streams}/server.sent.data/server/target"
+        "${server}/server.sent.data/server",
+        "${client}/server.sent.data/client"
     })
     public void shouldSpinWrite() throws Exception
     {
         ProcessDataHelper.fragmentWrites(generate(() -> 0).limit(WRITE_SPIN_COUNT - 1));
         HandleWriteHelper.fragmentWrites(generate(() -> 0));
-        shouldReceiveServerSentData("server data");
+        k3po.finish();
     }
 
     @Test
     @Specification({
         "${route}/server/controller",
-        "${streams}/server.sent.data/server/target"
+        "${server}/server.sent.data/server",
+        "${client}/server.sent.data/client"
     })
     public void shouldFinishWriteWhenSocketIsWritableAgain() throws Exception
     {
         ProcessDataHelper.fragmentWrites(IntStream.of(5));
-        shouldReceiveServerSentData("server data");
+        k3po.finish();
     }
 
     @Test
     @Specification({
         "${route}/server/controller",
-        "${streams}/server.sent.data/server/target"
+        "${server}/server.sent.data/server",
+        "${client}/server.sent.data/client"
     })
     public void shouldHandleMultiplePartialWrites() throws Exception
     {
         ProcessDataHelper.fragmentWrites(IntStream.of(2));
         HandleWriteHelper.fragmentWrites(IntStream.of(3, 1));
-        shouldReceiveServerSentData("server data");
+        k3po.finish();
     }
 
     @Test
     @Specification({
         "${route}/server/controller",
-        "${streams}/server.sent.data.multiple.frames/server/target"
+        "${server}/server.sent.data.multiple.frames/server",
+        "${client}/server.sent.data.multiple.frames/client"
     })
     public void shouldWriteWhenMoreDataArrivesWhileAwaitingSocketWritable() throws Exception
     {
@@ -132,13 +132,13 @@ public class ServerPartialWriteIT
         ProcessDataHelper.fragmentWrites(concat(of(5), generate(() -> finishWrite.getAndSet(true) ? 0 : 0)));
         HandleWriteHelper.fragmentWrites(generate(() -> finishWrite.get() ? ALL : 0));
 
-        shouldReceiveServerSentData("server data 1server data 2");
+        k3po.finish();
     }
 
     @Test
     @Specification({
         "${route}/server/controller",
-        "${streams}/server.sent.data.then.end/server/target"
+        "${server}/server.sent.data.then.end/server"
     })
     public void shouldHandleEndOfStreamWithPendingWrite() throws Exception
     {
@@ -178,98 +178,6 @@ public class ServerPartialWriteIT
             }
 
             assertTrue("Stream was not closed", closed);
-
-            k3po.finish();
-        }
-    }
-
-    @Test
-    @Specification({
-        "${route}/server/controller",
-        "${streams}/server.sent.data.after.end/server/target"
-    })
-    public void shouldResetIfDataReceivedAfterEndOfStreamWithPendingWrite() throws Exception
-    {
-        ProcessDataHelper.fragmentWrites(IntStream.of(6));
-        AtomicBoolean resetReceived = new AtomicBoolean(false);
-        HandleWriteHelper.fragmentWrites(generate(() -> resetReceived.get() ? ALL : 0));
-
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_SERVER");
-
-        try (SocketChannel channel = SocketChannel.open())
-        {
-            channel.connect(new InetSocketAddress("127.0.0.1", 0x1f90));
-
-            k3po.awaitBarrier("RESET_RECEIVED");
-            resetReceived.set(true);
-
-            ByteBuffer buf = ByteBuffer.allocate("server data".length() + 10);
-            boolean closed = false;
-            do
-            {
-                int len = channel.read(buf);
-                if (len == -1)
-                {
-                    closed = true;
-                    break;
-                }
-            } while (buf.position() < "server data".length());
-            buf.flip();
-
-            assertEquals("server data", UTF_8.decode(buf).toString());
-
-            if (!closed)
-            {
-                buf.rewind();
-                closed = (channel.read(buf) == -1);
-            }
-
-            assertTrue("Stream was not closed", closed);
-
-            k3po.finish();
-        }
-    }
-
-    private void shouldReceiveServerSentData(String expectedData) throws Exception
-    {
-        shouldReceiveServerSentData(expectedData, false);
-    }
-
-    private void shouldReceiveServerSentData(String expectedData, boolean expectStreamClosed) throws Exception
-    {
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_SERVER");
-
-        try (SocketChannel channel = SocketChannel.open())
-        {
-            channel.connect(new InetSocketAddress("127.0.0.1", 0x1f90));
-
-            ByteBuffer buf = ByteBuffer.allocate(expectedData.length() + 10);
-            boolean closed = false;
-            do
-            {
-                int len = channel.read(buf);
-                if (len == -1)
-                {
-                    closed = true;
-                    break;
-                }
-            } while (buf.position() < expectedData.length());
-            buf.flip();
-
-            assertEquals(expectedData, UTF_8.decode(buf).toString());
-
-            if (expectStreamClosed)
-            {
-                if (!closed)
-                {
-                    buf.rewind();
-                    closed = (channel.read(buf) == -1);
-                }
-
-                assertTrue("Stream was not closed", closed);
-            }
 
             k3po.finish();
         }
