+++ src/test/java/org/reaktivity/nukleus/tcp/internal/streams/rfc793/ServerResetAndAbortT.java
@@ -34,10 +34,10 @@ import org.junit.rules.Timeout;
 import org.junit.runner.RunWith;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper;
 import org.reaktivity.nukleus.tcp.internal.TcpController;
 import org.reaktivity.nukleus.tcp.internal.TcpCountersRule;
-import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper;
-import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.CountDownHelper;
+import org.reaktivity.nukleus.tcp.internal.SocketChannelHelper.CountDownHelper;
 import org.reaktivity.reaktor.test.ReaktorRule;
 
 /**
@@ -100,7 +100,7 @@ public class ServerResetAndAbortT
     @BMRule(name = "shutdownInput",
     targetClass = "^java.nio.channels.SocketChannel",
     targetMethod = "shutdownInput()",
-    helper = "org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper$CountDownHelper",
+    helper = "org.reaktivity.nukleus.tcp.internal.SocketChannelHelper$CountDownHelper",
     condition =
       "callerMatches(\"org.reaktivity.nukleus.tcp.internal.stream.ReadStream..*\", true, true)",
       action = "countDown()"
@@ -135,7 +135,7 @@ public class ServerResetAndAbortT
     @BMRule(name = "shutdownInput",
     targetClass = "^java.nio.channels.SocketChannel",
     targetMethod = "shutdownInput()",
-    helper = "org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper$CountDownHelper",
+    helper = "org.reaktivity.nukleus.tcp.internal.SocketChannelHelper$CountDownHelper",
     condition =
       "callerMatches(\"org.reaktivity.nukleus.tcp.internal.stream.ReadStream..*\", true, true)",
       action = "countDown()"
